Level 1 changes
- Added level 1-2, 1-3
- Added colliders on all objects across levels
- scaled prefabs and levels to increase the quality of the images

+++ Assets/Scripts/PlayerControl/PlayerMovement.cs
@@ -6,7 +6,7 @@ public class PlayerMovement : MonoBehaviour {
 	public bool playerIsdead= false;
 
 	// Use this for initialization
-	public int PlayerMovementspeed; 
+	public float PlayerMovementspeed; 
 
 	
 	// Update is called once per frame
